Add TypeFormatter.
This is needed to print function arguments for private symbols and inlined functions.
The arguments are not stored in a string, they're stored in the type data.

This code is based on @calixteman's TypeDumper from mozilla/dump_syms, which in turn
was probably based on the original Google breakpad code.
This code has undergone a number of changes while it was in the
mstange/profiler-get-symbols repo; if you want the full blame / history, you'll
need to check the history over there. You can start here:
https://github.com/mstange/profiler-get-symbols/blame/ee43039a9e85886dfcd19e4fa2717d3b2a29fab0/lib/src/windows/type_formatter.rs

The symbolic repo also has a TypeFormatter, but it has less functionality -
symbolic's TypeFormatter cannot dump function arguments.
The changes that I've made to TypeDumper brought its API closer to symbolic's
TypeFormatter API.

One missing piece of functionality is cross-module references.