Rewrite type size caching.
In the past we would parse all TypeDatas as we filed the type finder
cache, so that we could store the type's size. This parsing was quite
expensive and in many cases the sizes weren't needed.
Now we only parse types for storing their size if the size is requested.
We also keep track of the TypeIndex ranges that we've parsed already.