Get function arguments in more cases.
If we have both a public symbol and a procedure symbol for the same rva,
usually we want to prefer the procedure symbol because it has richer
information and is associated with its type index, which is how we get
to the function arguments.
However, if the procedure symbol is not associated with a type, then
we can't get the function arguments that way. In those scenarios, the
public symbol usually is the better alternative, because it contains
the arguments in a mangled form ('decorated').